tests/wpt/tests/html/dom/render-blocking/element-render-blocking-012.tentative.html

<!DOCTYPE html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/utils.js"></script>
<title>Href attribute changes in the head to apply</title>

<link id=link rel=expect href="" blocking="render">
<script>
link.href = "#last";

async_test((t) => {
  requestAnimationFrame(() => {
    t.step(() => assert_true(!!document.getElementById("last")));
    t.done();
  });
}, "adding href in the head makes it blocking");
</script>
</head>
<body>
  <div id="first"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="second"></div>
  <script>
    jankMany(100, 10);
  </script>
  <div id="last"></div>
</body>
